UserAdmin: Authentication: fix a bunch of typos and move validators in a separate folder
