+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Authentication/Header/AccountDataHeaderHeaderSerializer.cs
@@ -1,18 +1,18 @@
-﻿using Smart.FA.Catalog.Core.Extensions;
+using Smart.FA.Catalog.Core.Extensions;
 
 namespace Smart.FA.Catalog.Web.Authentication.Header;
 
 public class AccountDataHeaderHeaderSerializer : IAccountDataHeaderSerializer
 {
-    private AccountData CreateMock()
+    private AccountData CreateFakeCustomData()
     {
-        return new AccountData { FirstName = "Maxime", LastName = "P.", Email = "maxime@maxime.com" };
+        return new AccountData { FirstName = "John", LastName = "Doe", Email = "john.doe@unknown.com" };
     }
 
     /// <inheritdoc />
-    public string CreateSerializedMock()
+    public string CreateFakeAccountDataHeader()
     {
-        var customData = CreateMock();
+        var customData = CreateFakeCustomData();
         var serializedCustomData = customData.ToJson();
         return System.Web.HttpUtility.UrlDecode(serializedCustomData);
     }
