+++ src/UserAdmin/src/Smart.FA.Catalog.Web/Authentication/Header/IAccountDataHeaderSerializer.cs
@@ -1,15 +1,15 @@
-﻿namespace Smart.FA.Catalog.Web.Authentication.Header;
+namespace Smart.FA.Catalog.Web.Authentication.Header;
 
 public interface IAccountDataHeaderSerializer
 {
     /// <summary>
-    /// Create a mocked AccountData serialized in a json string
+    /// Creates a fake AccountData serialized in a json string.
     /// </summary>
     /// <returns>A json string with serialized AccountData</returns>
-    string CreateSerializedMock();
+    string CreateFakeAccountDataHeader();
 
     /// <summary>
-    /// Deserialize a json string into an AccountData object
+    /// Deserializes a json string into an AccountData object.
     /// </summary>
     /// <param name="serializedCustomData">The json string of an AccountData object (from the customData header)</param>
     /// <returns>An AccountData object</returns>
